Use standard mixer recipes for potion crafting
