+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/MixerTileEntity.java
@@ -17,7 +17,6 @@ import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.IBlockBou
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.IInteractionObjectIE;
 import blusunrize.immersiveengineering.common.blocks.generic.PoweredMultiblockTileEntity;
 import blusunrize.immersiveengineering.common.blocks.multiblocks.IEMultiblocks;
-import blusunrize.immersiveengineering.common.crafting.MixerRecipePotion;
 import blusunrize.immersiveengineering.common.util.CapabilityReference;
 import blusunrize.immersiveengineering.common.util.Utils;
 import blusunrize.immersiveengineering.common.util.inventory.IEInventoryHandler;
@@ -36,7 +35,6 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Vec3d;
 import net.minecraft.util.math.shapes.ISelectionContext;
 import net.minecraft.util.math.shapes.VoxelShape;
-import net.minecraftforge.common.brewing.BrewingRecipeRegistry;
 import net.minecraftforge.common.capabilities.Capability;
 import net.minecraftforge.common.util.Constants.NBT;
 import net.minecraftforge.common.util.LazyOptional;
@@ -512,21 +510,6 @@ public class MixerTileEntity extends PoweredMultiblockTileEntity<MixerTileEntity
 				}
 			}
 			super.doProcessTick(multiblock);
-		}
-
-		@Override
-		protected void processFinish(PoweredMultiblockTileEntity<?, MixerRecipe> multiblock)
-		{
-			super.processFinish(multiblock);
-			if(this.recipe instanceof MixerRecipePotion)
-				for(int i : this.inputSlots)
-					if(!multiblock.getInventory().get(i).isEmpty()&&
-							BrewingRecipeRegistry.isValidIngredient(multiblock.getInventory().get(i)))
-					{
-						multiblock.getInventory().get(i).shrink(1);
-						if(multiblock.getInventory().get(i).getCount() <= 0)
-							multiblock.getInventory().set(i, ItemStack.EMPTY);
-					}
 		}
 	}
