+++ src/main/java/blusunrize/immersiveengineering/common/crafting/serializers/MixerRecipeSerializer.java
@@ -14,8 +14,6 @@ import blusunrize.immersiveengineering.api.crafting.IERecipeSerializer;
 import blusunrize.immersiveengineering.api.crafting.IngredientWithSize;
 import blusunrize.immersiveengineering.api.crafting.MixerRecipe;
 import blusunrize.immersiveengineering.common.blocks.IEBlocks.Multiblocks;
-import blusunrize.immersiveengineering.common.crafting.MixerRecipePotion;
-import com.google.common.base.Preconditions;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import net.minecraft.item.ItemStack;
@@ -23,11 +21,8 @@ import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.JSONUtils;
 import net.minecraft.util.ResourceLocation;
 import net.minecraftforge.fluids.FluidStack;
-import org.apache.commons.lang3.tuple.Pair;
 
 import javax.annotation.Nullable;
-import java.util.ArrayList;
-import java.util.List;
 
 public class MixerRecipeSerializer extends IERecipeSerializer<MixerRecipe>
 {
@@ -55,62 +50,23 @@ public class MixerRecipeSerializer extends IERecipeSerializer<MixerRecipe>
 	public MixerRecipe read(ResourceLocation recipeId, PacketBuffer buffer)
 	{
 		FluidStack fluidOutput = buffer.readFluidStack();
-		final boolean isPotionRecipe = buffer.readBoolean();
-		int numInputs;
-		if(isPotionRecipe)
-			numInputs = buffer.readInt();
-		else
-			numInputs = 1;
-		List<Pair<FluidTagInput, IngredientWithSize[]>> inputs = new ArrayList<>(numInputs);
-		for(int i = 0; i < numInputs; ++i)
-		{
-			FluidTagInput fluidInput = FluidTagInput.read(buffer);
-			int ingredientCount = buffer.readInt();
-			IngredientWithSize[] itemInputs = new IngredientWithSize[ingredientCount];
-			for(int j = 0; j < ingredientCount; j++)
-				itemInputs[j] = IngredientWithSize.read(buffer);
-			inputs.add(Pair.of(fluidInput, itemInputs));
-		}
-		Pair<FluidTagInput, IngredientWithSize[]> firstInput = inputs.get(0);
-		if(isPotionRecipe)
-		{
-			MixerRecipePotion ret = new MixerRecipePotion(recipeId, fluidOutput, firstInput.getLeft(), firstInput.getRight());
-			for(int i = 1; i < inputs.size(); ++i)
-			{
-				Pair<FluidTagInput, IngredientWithSize[]> input = inputs.get(i);
-				Preconditions.checkState(input.getRight().length==1);
-				ret.addAlternateInput(input.getLeft(), input.getRight()[0]);
-			}
-			return ret;
-		}
-		else
-		{
-			int energy = buffer.readInt();
-			return new MixerRecipe(recipeId, fluidOutput, firstInput.getLeft(), firstInput.getRight(), energy);
-		}
+		FluidTagInput fluidInput = FluidTagInput.read(buffer);
+		int ingredientCount = buffer.readInt();
+		IngredientWithSize[] itemInputs = new IngredientWithSize[ingredientCount];
+		for(int j = 0; j < ingredientCount; j++)
+			itemInputs[j] = IngredientWithSize.read(buffer);
+		int energy = buffer.readInt();
+		return new MixerRecipe(recipeId, fluidOutput, fluidInput, itemInputs, energy);
 	}
 
 	@Override
 	public void write(PacketBuffer buffer, MixerRecipe recipe)
 	{
 		buffer.writeFluidStack(recipe.fluidOutput);
-		buffer.writeBoolean(recipe instanceof MixerRecipePotion);
-		Pair<FluidTagInput, IngredientWithSize[]> firstInput = Pair.of(recipe.fluidInput, recipe.itemInputs);
-		List<Pair<FluidTagInput, IngredientWithSize[]>> inputs = new ArrayList<>();
-		inputs.add(firstInput);
-		if(recipe instanceof MixerRecipePotion)
-		{
-			inputs.addAll(((MixerRecipePotion)recipe).getAlternateInputs());
-			buffer.writeInt(inputs.size());
-		}
-		for(Pair<FluidTagInput, IngredientWithSize[]> inputPair : inputs)
-		{
-			inputPair.getLeft().write(buffer);
-			buffer.writeInt(inputPair.getRight().length);
-			for(IngredientWithSize input : inputPair.getRight())
-				input.write(buffer);
-		}
-		if(!(recipe instanceof MixerRecipePotion))
-			buffer.writeInt(recipe.getTotalProcessEnergy());
+		recipe.fluidInput.write(buffer);
+		buffer.writeInt(recipe.itemInputs.length);
+		for(IngredientWithSize input : recipe.itemInputs)
+			input.write(buffer);
+		buffer.writeInt(recipe.getTotalProcessEnergy());
 	}
 }
